requirements.txt
django==4.2.4
djangorestframework==3.14.0
django-environ==0.10.0
django-filter==23.2
requests==2.31.0
psycopg2-binary==2.9.7